style: finish css for table and fix a bug where settings.properties would be always created

+++ Jayuwoki/src/main/java/dad/utils/Utils.java
@@ -1,5 +1,6 @@
 package dad.utils;
 
+import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
@@ -29,17 +30,20 @@ public class Utils {
     public static void setTheme(String theme) {
         properties.setProperty("theme", theme);
         saveProperties();
+        System.out.println("Theme saved: " + theme); // Debugging line
     }
 
     public static void createPropertiesFile(String filePath) {
-        Properties properties = new Properties();
-        try (FileOutputStream output = new FileOutputStream(filePath)) {
-            properties.setProperty("theme", "dark");
-            properties.setProperty("language", "en");
-            // Save properties to the specified file
-            properties.store(output, null);
-        } catch (IOException e) {
-            e.printStackTrace();
+        File file = new File(filePath);
+        if (!file.exists()) {
+            Properties properties = new Properties();
+            try (FileOutputStream output = new FileOutputStream(filePath)) {
+                properties.setProperty("theme", "dark");
+                properties.setProperty("language", "en");
+                properties.store(output, null);
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
         }
     }
 }
